refactor: CertificationController에 파사드 패턴 적용
- 파사드 패턴 적용에 따라 인터페이스 추가 및 구현체 적용
- 이름 변경
