+++ src/main/java/com/genius/gitget/challenge/certification/controller/CertificationController.java
@@ -8,7 +8,7 @@ import com.genius.gitget.challenge.certification.dto.CertificationResponse;
 import com.genius.gitget.challenge.certification.dto.InstancePreviewResponse;
 import com.genius.gitget.challenge.certification.dto.TotalResponse;
 import com.genius.gitget.challenge.certification.dto.WeekResponse;
-import com.genius.gitget.challenge.certification.service.CertificationService;
+import com.genius.gitget.challenge.certification.facade.CertificationFacade;
 import com.genius.gitget.challenge.certification.util.DateUtil;
 import com.genius.gitget.challenge.instance.domain.Instance;
 import com.genius.gitget.challenge.instance.service.InstanceProvider;
@@ -43,7 +43,7 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/certification")
 public class CertificationController {
     private final UserService userService;
-    private final CertificationService certificationService;
+    private final CertificationFacade certificationFacade;
     private final InstanceProvider instanceProvider;
     private final ParticipantService participantService;
 
@@ -52,7 +52,7 @@ public class CertificationController {
     public ResponseEntity<SingleResponse<InstancePreviewResponse>> getInstanceInformation(
             @PathVariable Long instanceId
     ) {
-        InstancePreviewResponse instancePreviewResponse = certificationService.getInstancePreview(instanceId);
+        InstancePreviewResponse instancePreviewResponse = certificationFacade.getInstancePreview(instanceId);
 
         return ResponseEntity.ok().body(
                 new SingleResponse<>(SUCCESS.getStatus(), SUCCESS.getMessage(), instancePreviewResponse)
@@ -64,7 +64,7 @@ public class CertificationController {
             @AuthenticationPrincipal UserPrincipal userPrincipal,
             @RequestBody CertificationRequest certificationRequest
     ) {
-        CertificationResponse certificationResponse = certificationService.updateCertification(
+        CertificationResponse certificationResponse = certificationFacade.updateCertification(
                 userPrincipal.getUser(), certificationRequest);
 
         return ResponseEntity.ok().body(
@@ -78,7 +78,7 @@ public class CertificationController {
             @RequestBody CertificationRequest certificationRequest
     ) {
         User user = userPrincipal.getUser();
-        ActivatedResponse activatedResponse = certificationService.passCertification(
+        ActivatedResponse activatedResponse = certificationFacade.passCertification(
                 user.getId(), certificationRequest);
 
         return ResponseEntity.ok().body(
@@ -93,7 +93,7 @@ public class CertificationController {
     ) {
         LocalDate kstDate = DateUtil.convertToKST(LocalDateTime.now());
         Participant participant = participantService.findByJoinInfo(userPrincipal.getUser().getId(), instanceId);
-        WeekResponse weekResponse = certificationService.getMyWeekCertifications(participant.getId(), kstDate);
+        WeekResponse weekResponse = certificationFacade.getMyWeekCertifications(participant.getId(), kstDate);
 
         return ResponseEntity.ok().body(
                 new SingleResponse<>(SUCCESS.getStatus(), SUCCESS.getMessage(), weekResponse)
@@ -108,7 +108,7 @@ public class CertificationController {
     ) {
         LocalDate kstDate = DateUtil.convertToKST(LocalDateTime.now());
         User user = userPrincipal.getUser();
-        Slice<WeekResponse> certifications = certificationService.getOthersWeekCertifications(
+        Slice<WeekResponse> certifications = certificationFacade.getOthersWeekCertifications(
                 user.getId(), instanceId, kstDate, pageable);
 
         return ResponseEntity.ok().body(
@@ -124,7 +124,7 @@ public class CertificationController {
         LocalDate kstDate = DateUtil.convertToKST(LocalDateTime.now());
         User user = userService.findUserById(userId);
         Participant participant = participantService.findByJoinInfo(user.getId(), instanceId);
-        TotalResponse totalResponse = certificationService.getTotalCertification(
+        TotalResponse totalResponse = certificationFacade.getTotalCertification(
                 participant.getId(), kstDate);
 
         return ResponseEntity.ok().body(
@@ -144,7 +144,7 @@ public class CertificationController {
                 userPrincipal.getUser().getId(),
                 instanceId);
 
-        CertificationInformation certificationInformation = certificationService.getCertificationInformation(
+        CertificationInformation certificationInformation = certificationFacade.getCertificationInformation(
                 instance, participant, kstDate);
 
         return ResponseEntity.ok().body(
